circular/requirements.txt
Django==1.6.2
Markdown==2.4
Pillow==2.4.0
South==0.8.4
dj-database-url==0.3.0
dj-static==0.0.5
django-filter==0.7
django-registration==1.0
djangorestframework==2.3.13
gunicorn==18.0
psycopg2==2.5.2
pylint==1.2.1
pystache==0.5.3
requests==2.2.1
static==1.0.2
wsgiref==0.1.2
yummly==0.4.0
